+++ backend/src/db/schema.ts
@@ -1,23 +1,8 @@
-import { pgTable, uuid, varchar, timestamp } from 'drizzle-orm/pg-core';
-import { z } from 'zod';
+// Re-export all schemas for backward compatibility
+export * from './schema/index';
 
-// User table schema
-export const users = pgTable('users', {
-  id: uuid('id').primaryKey().defaultRandom(),
-  name: varchar('name', { length: 100 }).notNull(),
-  email: varchar('email', { length: 255 }).notNull().unique(),
-  password: varchar('password', { length: 255 }).notNull(),
-  createdAt: timestamp('created_at', { withTimezone: true }).defaultNow().notNull(),
-  updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow().notNull(),
-});
+// Re-export validation schemas for backward compatibility
+export { userValidationSchema, registerSchema, loginSchema } from '../validation/users';
 
-// Zod schemas for validation
-export const userValidationSchema = z.object({
-  name: z.string().min(1).max(100),
-  email: z.string().email(),
-  password: z.string().min(6),
-});
-
-// Types
-export type User = typeof users.$inferSelect;
-export type NewUser = typeof users.$inferInsert;
+// Re-export types for backward compatibility
+export type { User, NewUser, PublicUser } from '../types/users';
