+++ backend/src/db/schema/index.ts
@@ -1,0 +1,5 @@
+// Re-export all schemas for easy importing
+export * from './users';
+// Future exports:
+// export * from './posts';
+// export * from './comments';
