+++ backend/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './users';
+// Future exports:
+// export * from './posts';
+// export * from './auth';
